+++ libft/libft.h
@@ -69,7 +69,7 @@ char			*ft_strtrim(char const *s);
 char			**ft_strsplit(char const *s, char c);
 size_t			ft_strchlen(const char *s, char c);
 char			*ft_itoa(int n);
-char			*ft_itoa_base(unsigned long long int n, int base);
+char			*ft_itoa_base(__int128 n, int base);
 void			ft_putchar(char c);
 void			ft_putstr(char const *s);
 void			ft_putendl(char const *s);
